pyproject.toml
[tool.poetry]
name = "trade-quotes-be"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
strawberry-graphql = "^0.168.1"
beanie = "^1.17.0"
pyjwt = "^2.6.0"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
